Add processes state view in observer
State of generic processes is shown in a new tab of the existing
processes view.
Behaviour is also shown when declared into the module attributes,
otherwise 'undefined'.
If state format is not detected, Format is set to 'overriden' when
format_status/2 is defined, otherwise 'unknown'.
Some processes don't treat system messages, this imply a short delay
 for displaying the process view, arbitrary set to 200ms.